pyproject.toml
[tool.poetry]
name = "empatoes"
version = "0.1.0"
description = ""
authors = ["Davies <daviesaraujo@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pyinstaller = "^5.13.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
